+++ subprojects/messageutil/src/main/kotlin/osdntogit/ReformatMessage.kt
@@ -1,4 +1,4 @@
-package osdnto3
+package osdntogit
 
 import messageutil.Growth
 import messageutil.Love
