+++ subprojects/messageutil/src/main/kotlin/osdntogit/Write.kt
@@ -1,4 +1,4 @@
-package osdnto3
+package osdntogit
 
 import messageutil.myYaml
 import messageutil.renameMessages
